requirements.txt
asgiref==3.3.4
astroid==2.0.4
boto3==1.18.44
botocore==1.21.44
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.8
decorator==5.1.0
decouple==0.0.7
Django==3.2.3
django-cleanup==5.2.0
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-pagination==1.0.7
django-rest-knox==4.1.0
django-rest-passwordreset==1.2.0
django-shopping-cart==0.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
dnspython==2.1.0
dynaconf==3.1.7
Eel==0.14.0
future==0.18.2
geojson==2.5.0
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.1.0
hvac==0.11.0
idna==2.10
isort==4.3.4
Jinja2==3.0.1
jmespath==0.10.0
knox==0.1.14
lazy-object-proxy==1.3.1
loguru==0.5.3
MarkupSafe==2.0.1
mccabe==0.6.1
Pillow==8.2.0
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.4
pyowm==3.2.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2018.5
requests==2.25.1
s3transfer==0.5.0
shortuuid==1.0.1
six==1.11.0
sqlparse==0.4.1
urllib3==1.26.5
validators==0.18.2
whichcraft==0.6.1
win32-setctime==1.0.3
wrapt==1.10.11
xlwt==1.3.0
zope.event==4.5.0
zope.interface==5.4.0
